Refactor date range selection for Contributions into separate component
