+++ src/client/components/Pages/Auth/Contributions/CalendarGrid/CalendarGrid.module.css
@@ -2,7 +2,7 @@
     display: flex;
     flex-direction: column;
     flex-grow: 0;
-    width: 100%;
+    /* width: 100%; */
     max-width: 999px;
     overflow-x: auto;
     padding: 0 12px 12px 0;
