+++ src/client/components/Reuse/DateRangeInput/index.ts
@@ -1,0 +1,2 @@
+import { DateRangeInput } from './DateRangeInput'
+export default DateRangeInput
